ADD DTO AND CHANGE WAGON TYPE

+++ src/main/java/com/team/passengerrailwaytransportation/entities/Ticket.java
@@ -31,5 +31,5 @@ public class Ticket extends Domain {
   @OneToOne
   Transportation transportation;
 
-  BigInteger price;
+  Integer price;
 }
